services/user-api/requirements.txt
flask==3.1.1
flask-cors==5.0.1
pyjwt==2.10.1
python-dotenv==1.1.0
gunicorn==23.0.0
pytest==8.3.5
flake8==7.2.0
